requirements.txt
Flask==2.3.2
pandas==1.5.3
numpy==1.26.4
fpdf==1.7.2
Werkzeug==2.3.6
python-dotenv==1.0.1
gunicorn==23.0.0